+++ admin/src/main/java/org/yapp/auth/presentation/AuthController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.yapp.auth.AuthToken;
 import org.yapp.auth.application.AuthService;
 import org.yapp.auth.presentation.request.LoginRequest;
-import org.yapp.util.CommonResponse;
+import org.yapp.format.CommonResponse;
 
 @RestController
 @RequiredArgsConstructor
