+++ admin/src/main/java/org/yapp/report/application/ReportService.java
@@ -8,11 +8,11 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.yapp.domain.report.Report;
+import org.yapp.format.PageResponse;
 import org.yapp.report.application.dto.ReportedUserWithReasonDto;
 import org.yapp.report.dao.ReportRepository;
 import org.yapp.report.presentation.response.ReportDetailResponse;
 import org.yapp.report.presentation.response.ReportUserResponse;
-import org.yapp.util.PageResponse;
 
 @Service
 @RequiredArgsConstructor
