+++ api/src/main/java/org/yapp/domain/auth/presentation/LoginController.java
@@ -1,5 +1,7 @@
 package org.yapp.domain.auth.presentation;
 
+import io.swagger.v3.oas.annotations.Operation;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -9,31 +11,30 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.yapp.domain.auth.application.oauth.service.OauthService;
 import org.yapp.domain.auth.presentation.dto.request.OauthLoginRequest;
 import org.yapp.domain.auth.presentation.dto.response.OauthLoginResponse;
-import org.yapp.util.CommonResponse;
-
-import io.swagger.v3.oas.annotations.Operation;
-import lombok.RequiredArgsConstructor;
+import org.yapp.format.CommonResponse;
 
 @Controller
 @RequestMapping("/api/login")
 @RequiredArgsConstructor
 public class LoginController {
-  private final OauthService oauthService;
 
-  /**
-   * 개발중 소셜 로그인으로 사용자의 accessToken을 가져오기 어렵기 때문에 만든 임시 메서드
-   */
-  @PostMapping("/test/users/{userId}")
-  @Operation(summary = "테스트 로그인", description = "테스트용 임시 토큰을 발급합니다.", tags = {"로그인"})
-  public ResponseEntity<CommonResponse<OauthLoginResponse>> getToken(@PathVariable Long userId) {
-    OauthLoginResponse response = oauthService.tmpTokenGet(userId);
-    return ResponseEntity.ok(CommonResponse.createSuccess(response));
-  }
+    private final OauthService oauthService;
 
-  @PostMapping("/oauth")
-  @Operation(summary = "Oauth 로그인", description = "Oauth 로그인을 하고 토큰들을 발급합니다.", tags = {"로그인"})
-  public ResponseEntity<CommonResponse<OauthLoginResponse>> oauthLogin(@RequestBody OauthLoginRequest request) {
-    OauthLoginResponse response = oauthService.login(request);
-    return ResponseEntity.ok(CommonResponse.createSuccess(response));
-  }
+    /**
+     * 개발중 소셜 로그인으로 사용자의 accessToken을 가져오기 어렵기 때문에 만든 임시 메서드
+     */
+    @PostMapping("/test/users/{userId}")
+    @Operation(summary = "테스트 로그인", description = "테스트용 임시 토큰을 발급합니다.", tags = {"로그인"})
+    public ResponseEntity<CommonResponse<OauthLoginResponse>> getToken(@PathVariable Long userId) {
+        OauthLoginResponse response = oauthService.tmpTokenGet(userId);
+        return ResponseEntity.ok(CommonResponse.createSuccess(response));
+    }
+
+    @PostMapping("/oauth")
+    @Operation(summary = "Oauth 로그인", description = "Oauth 로그인을 하고 토큰들을 발급합니다.", tags = {"로그인"})
+    public ResponseEntity<CommonResponse<OauthLoginResponse>> oauthLogin(
+        @RequestBody OauthLoginRequest request) {
+        OauthLoginResponse response = oauthService.login(request);
+        return ResponseEntity.ok(CommonResponse.createSuccess(response));
+    }
 }
