+++ api/src/main/java/org/yapp/domain/term/presentation/TermController.java
@@ -2,6 +2,7 @@ package org.yapp.domain.term.presentation;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,14 +18,13 @@ import org.yapp.domain.term.applicatoin.TermUseCase;
 import org.yapp.domain.term.applicatoin.dto.SignupTermsDto;
 import org.yapp.domain.term.presentation.dto.request.TermAgreementRequest;
 import org.yapp.domain.term.presentation.dto.response.TermResponses;
-import org.yapp.util.CommonResponse;
-
-import java.util.List;
+import org.yapp.format.CommonResponse;
 
 @RequestMapping("/api/terms")
 @Controller
 @RequiredArgsConstructor
 public class TermController {
+
     private final TermService termService;
     private final TermUseCase termUseCase;
 
@@ -34,16 +34,16 @@ public class TermController {
     public ResponseEntity<CommonResponse<TermResponses>> getAllTerms() {
         List<Term> allActiveTerms = termService.getAllActiveTerms();
         return ResponseEntity.status(HttpStatus.OK)
-                .body(CommonResponse.createSuccess(TermResponses.from(allActiveTerms)));
+            .body(CommonResponse.createSuccess(TermResponses.from(allActiveTerms)));
     }
 
     @PostMapping("/agree")
     @Operation(summary = "사용자 약관 동의", description = "사용자가 설정한 약관을 등록합니다", tags = {"약관"})
     @ApiResponse(responseCode = "200", description = "사용자가 설정한 약관이 등록되었습니다.")
     public ResponseEntity<CommonResponse<Void>> agreeTerm(@AuthenticationPrincipal Long userId,
-                                                          @RequestBody TermAgreementRequest request) {
+        @RequestBody TermAgreementRequest request) {
         termUseCase.checkTermConstraints(new SignupTermsDto(userId, request.agreedTermsId()));
         return ResponseEntity.status(HttpStatus.OK)
-                .body(CommonResponse.createSuccessWithNoContent());
+            .body(CommonResponse.createSuccessWithNoContent());
     }
 }
