+++ common/format/src/main/java/org/yapp/format/PageResponse.java
@@ -1,4 +1,4 @@
-package org.yapp.util;
+package org.yapp.format;
 
 import java.util.List;
 import lombok.AllArgsConstructor;
